Add CodeArea tabs and syntax highlighting
RichTextFX doesn't count /r characters for some reason so we're stuck
with some ugly code to work around that. I also had to remove the
PsiCommentElement in order to make this work, but I haven't tried ading
it back or something so I'll try that later and see if I can make that
work again.